Move Base45Encoder to appropriate package
