+++ src/commonTest/kotlin/ehn/techiop/hcert/kotlin/chain/impl/Base45EncoderTest.kt
@@ -1,6 +1,5 @@
 package ehn.techiop.hcert.kotlin.chain.impl
 
-import ehn.techiop.hcert.kotlin.chain.common.Base45Encoder
 import ehn.techiop.hcert.kotlin.chain.fromHexString
 import ehn.techiop.hcert.kotlin.chain.toHexString
 import io.kotest.assertions.throwables.shouldThrow
