+++ src/jvmMain/datagen/ehn/techiop/hcert/kotlin/chain/faults/FaultyBase45Service.kt
@@ -1,6 +1,6 @@
 package ehn.techiop.hcert.kotlin.chain.faults
 
-import ehn.techiop.hcert.kotlin.chain.common.Base45Encoder
+import ehn.techiop.hcert.kotlin.chain.impl.Base45Encoder
 import ehn.techiop.hcert.kotlin.chain.impl.DefaultBase45Service
 
 /**
